scraper/pyproject.toml
[tool.poetry]
name = "scraper"
version = "0.1.0"
description = ""
authors = ["StAandrew <dev.staniukynas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
firecrawl-py = "^1.3.1"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
